+++ log/context_test.go
@@ -90,3 +90,37 @@ func TestWithNamesRaceRegression(t *testing.T) {
 		}
 	})
 }
+
+func TestFieldsFromContext(t *testing.T) {
+	for _, tt := range []struct {
+		ctx    context.Context //nolint:containedctx
+		fields FieldsType
+	}{
+		{
+			ctx:    context.Background(),
+			fields: FieldsType{},
+		},
+		{
+			ctx:    WithFields(context.Background(), FieldsType{"a": "1", "b": "1"}),
+			fields: FieldsType{"a": "1", "b": "1"},
+		},
+		{
+			ctx: WithFields(
+				WithFields(context.Background(), FieldsType{"a": "1", "b": "1"}),
+				FieldsType{"a": "2", "b": "2"},
+			),
+			fields: FieldsType{"a": "2", "b": "2"},
+		},
+		{
+			ctx: WithFields(
+				WithFields(context.Background(), FieldsType{"a": "1", "b": "1"}),
+				FieldsType{"a": "3"},
+			),
+			fields: FieldsType{"a": "3", "b": "1"},
+		},
+	} {
+		t.Run("", func(t *testing.T) {
+			require.Equal(t, tt.fields, FieldsFromContext(tt.ctx))
+		})
+	}
+}